+++ common/core/src/main/java/is/codion/common/value/AbstractValue.java
@@ -103,7 +103,9 @@ public abstract class AbstractValue<T> implements Value<T> {
 
 	@Override
 	public final @Nullable T get() {
-		return getValue();
+		T value = getValue();
+
+		return value == null ? nullValue : value;
 	}
 
 	@Override